Testy pro vylepseny taskmanager/requirements.txt
mysql-connector-python
pytest
PyMySQ
dotenv